New sin wave map generation. Easier tile size switching. Better player death and input handling.

+++ Game/Projectile.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "math.h"
 #include "Globals.h"
 #include "TileMap.h"
 
@@ -17,5 +16,5 @@ private:
     sf::Vector2f velocity;
     sf::Vector2i tile_coords;
     bool is_expired;
-    int blast_radius;
+    float blast_radius;
 };
